[FE] rename: game 페이지 경로 수정
