+++ frontend/src/index.js
@@ -2,7 +2,7 @@ import Profile from '@pages/Profile.js';
 import Login from '@pages/Login.js';
 import Home from '@pages/Home.js';
 import Friends from '@pages/Friends.js';
-import Game from '@pages/Game.js';
+import Game from '@pages/game/Game.js';
 import NavBar from '@component/navigation/NavBar.js';
 import drawBackground from '@/background/background.js';
